Cutsom logo will fit well on mobile
Make sure the dashboard text fits and aligns next to custom logo's as well.
flexbox is used to vertically align the <span> holding the word Dashboard or its translation.
Because flex elements can not be aligned using text-align, margin auto is applied instead to center the logo horizontally. The margin is not applied on the <img> directly, but on all children of .logo, so it always allies, even on custom logo's.